Add clear.js file and import it to index.js
